// src/modules/programs/domain/enums/degree-level.enum.ts
export enum DegreeLevel {
  DIPLOMA = 'DIPLOMA',                          // دبلوم
  HIGHER_DIPLOMA = 'HIGHER_DIPLOMA',            // دبلوم عالي
  BACHELOR = 'BACHELOR',                        // بكالوريوس
  MASTER = 'MASTER',                            // ماجستير
  PHD = 'PHD',                                  // دكتوراه
}